requirements.txt
absl-py==1.2.0
asttokens==2.0.8
backcall==0.2.0
blessed==1.19.1
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cuda-python==11.8.0
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
docker-pycreds==0.4.0
executing==1.1.1
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.28
google-auth==2.12.0
google-auth-oauthlib==0.4.6
gpustat==1.0.0
grpcio==1.49.1
idna==3.4
importlib-metadata==5.0.0
ipython==8.5.0
jedi==0.18.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
natsort==8.2.0
numpy==1.23.3
nvidia-ml-py==11.495.46
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.2
seaborn==0.12.0
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor-whl==1.1.2
thop==0.1.1.post2209072238
tk==0.1.0
torch==1.11.0+cu115
torchaudio==0.11.0+cu115
torchvision==0.12.0+cu115
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.12.10
wcwidth==0.2.5
Werkzeug==2.2.2
yaspin==2.2.0
zipp==3.9.0
